requirements.txt
playwright==1.40.0
supabase==2.0.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3 